Implement "files" directive in config files
This adds a special case for a config file that acts as a reset
switch, by having a `files: ` entry instead of `patchUpdated` or
`commandUpdated`.

Since the desired behaviour is to act as though there were _no_ config
file in effect, by far the most direct implementation is to add the
path in question to the collection of raw files paths when a config
file containing the directive is found.